refactor(network host): split NetworkInspectorScreen into focused files
Move the traffic view, JSON body viewer and mocks view out of the single large
file (no behavior change):
- NetworkInspectorScreen.kt: the screen entry + mock-from-transaction builder
- TrafficView.kt: traffic list, row, status badge, transaction detail, headers
- JsonBodyView.kt: JSON tree / pretty-text body rendering
- MocksView.kt: mocks tab, rule card, HTTP method dropdown

Cross-file composables (TrafficTab/MocksTab/BodyBlock) are now internal.
